// angular-src/src/app/models/centralstationdatacms.ts
export class CentralStationDataCMS {
    id: number;
    loadDate: Date;
    site_no: string;
    system_no: string;
    site_name: string;
    site_addr1: string;
    street_no: string;
    street_name: string;
    sort_key: string;
    co_no: string;
    branch_no: string;
    sitetype_id: string;
    sitestat_id: string;
    country_name: string;
    timezone_no: string;
    geocode_id: string;
    taxauth_no: string;
    geo_ovr_flag: string;
    local_flag: string;
    ext_warr_flag: string;
    owner_occupy_flag: string;
    change_date: string;
    change_user: string;
    change_no: string;
    site_addr2: string;
    city_name: string;
    state_id: string;
    zip_code: string;
    county_name: string;
    phone1: string;
    ext1: string;
    phone2: string;
    ext2: string;
    cross_street: string;
    servarea_id: string;
    dst_no: string;
    grdpart_no: string;
    servpart_no: string;
    cspart_no: string;
    sitebstat_id: string;
    aq_no: string;
    mktsrc_id: string;
    ulcode_id: string;
    compet_id: string;
    sales_emp_no: string;
    codeword1: string;
    mapbook_id: string;
    map_page: string;
    map_coord: string;
    orig_install_date: string;
    lead_no: string;
    subdivision: string;
    owner_no: string;
    udf1: string;
    udf2: string;
    udf3: string;
    udf4: string;
    terr_id: string;
    sic_id: string;
    occ_id: string;
    phone1_rev: string;
    phone2_rev: string;
    prospect_no: string;
    corpacct_id: string;
    siteloc_id: string;
    install_servco_no: string;
    servco_no: string;
    corpacct_servco_no: string;
    key_no: string;
    service_comment: string;
    attention: string;
    recur_costctr_id: string;
    install_costctr_id: string;
    service_costctr_id: string;
    other_costctr_id: string;
    in_city_flag: string;
    route_id: string;
    position: string;
    patrol_servarea_id: string;
    patrol_servplan_id: string;
    zipdef_no: string;
    service_emp_no: string;
    longitude: string;
    latitude: string;
    telco_lease_line: string;
    cf_acdType: string;
    cs_no: string;
    ooscat_id: string;
    oos_start_date: string;
    systype_id: string;
    descr: string;
}