// angular-src/src/app/models/eventhistorydate.ts
export class EventHistoryDate {
    cS_Number: string;
    event_Date: string;
    seqno: string;
    cS_Event_No: string;
    event_Id: string    ;
    event_Class: string ;
    event: string;
    event_Rpt: string;
    zone_Id:  string;
    user_Id:  string;
    userName: string;
    operator: string;
    additional_Info: string;
    scheduled_Date: string;
    comment: string;
    alarm_Incident_No: string;
    operator_Dispatch_Id: string;
    pascom: string;
    server_Id: string;
    site_name: string;
    udf1: string;
    udf2: string;
    udf3: string;
    udf4: string;
    phone: string
}